requirements.txt
pytest
testcontainers[nats]
nats-py
pytest-asyncio

